fairsynthesis_data_model/pyproject.toml
[project]
name = "fairsynthesis-data-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Felix Neubauer"}
]
requires-python = ">=3.13"
dependencies = [
    "jsonschema>=4.24.0",
    "lxml>=5.4.0",
    "pubchempy>=1.0.4",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "sympy>=1.14.0",
]

[project.scripts]
fairsynthesis-data-model = "fairsynthesis_data_model:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
